requirements.txt
langchain
langchain-community
PyPDF2
python-dotenv
streamlit
openai
faiss-cpu
altair
tiktoken
# uncomment to use instructor embeddings
InstructorEmbedding
sentence-transformers
beautifulsoup4
pandas